feat(tabs): implement default tabs and improve TabsProperty component logic
